refactor compute loop strategy
