+++ crates/cubecl-linalg/src/matmul/cmma/compute_loop/mod.rs
@@ -1,1 +1,3 @@
+mod accumulators_first;
 pub(crate) mod base;
+mod buffers_first;
